Yet another dependency/database configuration problem as surfaced:
Either of the following pseudo-call stacks illustrates the problem:

    1) configure models -> load & apply config -> reinstall addins -> reconfigure models
      or even:
    2) configure models -> ensure addin installation -> reconfigure models

In certain situation this will lead to db objects from a previous config to be used together from already reconfigured objects, which breaks terribly.

It turns out the core reason for this is that currently, much of the module-interdependencies is sort a mess, with no clear policies as to what changes propagate to where.

This commit attempts to establish some rules and policies for this, as well as fix the above problem, the latter by basically only updating module dependencies in one direction, and requiring the user to do the other direction manually, should it be necessary.